+++ app.py
@@ -3,8 +3,8 @@ import logging.config
 import os
 from flask import Flask, Blueprint
 import settings
-from lonelyfoodie.api.controllers.controller import ns as restaurants_namespace
-from lonelyfoodie.api.controllers.controller import ns_user as user_namespace
+from lonelyfoodie.api.controllers.restaurant_controller import ns as restaurants_namespace
+from lonelyfoodie.api.controllers.user_controller import ns_user as user_namespace
 from lonelyfoodie.api.restx import api
 
 app = Flask(__name__)
